requirements.txt
fastapi
uvicorn
requests
python-dotenv
Pillow
python-multipart